requirements.txt
streamlit 
pypdf 
python-docx 
transformers
sentencepiece
PyPDF2